Move test-infra-definitions inside datadog-agent part 1 (#43238)
This PR is the first part of that global PR: https://github.com/DataDog/datadog-agent/pull/41937 that was split to make it easier and safer to review.

Move test-infra-definitions inside datadog-agent, under `test/e2e-framework`. The content of `test/new-e2e/pkg` is also moved inside the `test/e2e-framework` module.

https://docs.google.com/document/d/1ML5R_tmFz8K7RiC5TOCql4dve3PLpGf1GZEVS-mkaUA/edit?tab=t.0#heading=h.ixldy9ty2xy5

Pipeline running should be entirely green.

I will handle eventual drift between test-infra-definitions on the separate repository and that current version, for the PR that are merged in the meantime

Note for reviewer: You can review only the files you own, the changes made to the test should not change anything about how the test is configured. Mostly wrapping the existing thing inside `WithRunOptions`

Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>