[CWS] Revert "Re-apply "[CWS] SECL variables refactoring"" (#43398)
### What does this PR do?

Reverts PR https://github.com/DataDog/datadog-agent/pull/42315 as well as https://github.com/DataDog/datadog-agent/pull/43284 that depends on it. 

### Motivation

The reverted PR introduced a regression in the way inherited variables are being set.
This regression was not caught by existing tests so we are reverting the PR for now, and will reintroduce it after we added new tests that would have caught the regression. 

### Describe how you validated your changes

### Additional Notes

#incident-46164


Co-authored-by: yoann.ghigoff <yoann.ghigoff@datadoghq.com>