Replace Viper work-arounds in OTel's config with new methods (#43204)
### What does this PR do?

Replace Viper work-arounds in OTel's config code with new methods from the config: `HasSection` and `GetViperCombine`.

`HasSection` returns true for config "sections", which look like this:

```
section:
  label:
```

whether or not anything appears below `label`

`GetViperCombine` is a work-around for Viper not merging layers when called like `cfg.Get(key)`. It was originally used by `UnmarshalKey` but works here as well. Generally, `GetViperCombine` can't be used as a drop-in replacement for `cfg.Get` because it has much worse performance. Furthermore, OTel's config code can't use `UnmarshalKey` because that method doesn't retain empty sections.

This PR is best reviewed on commit at a time.

### Motivation

Code shouldn't rely on `AllSettings()` or `AllKeysLowerCased()` as they have unclear semantics around what exactly they return. It's better to have methods in the config do the heavy lifting.

### Describe how you validated your changes

Unit tests cover the basic functions `HasSection` and `GetViperCombine` for all the different cases we care about. The existing otelcol tests all pass, with only one minor change which is needed because this PR removes the hard-coded list of which settings need to be cast to int. Instead, the code now relies upon Viper's behavior: if a setting has a default value, then an env var of that value can be cast to the same type.

### Additional Notes


Co-authored-by: dustin.long <dustin.long@datadoghq.com>