[kubernetes/kubelet] Disable pod list cache by default (#43333)
### What does this PR do?

This PR disables the pod list cache in the Kubelet client by default.

The only part of the code that lists pods from the kubelet is workloadmeta, and it already has its own storage. The frequency at which workloadmeta pulls from the kubelet isn’t just set by its own pull interval, we also have to consider the cache TTL, which was set to 5s by default.

In short, keeping another cache in the Kubelet client duplicates data, makes debugging harder, and adds delay. Disabling it may increase the pull frequency, but the impact should be minimal, and having more up to date data is probably worth it.

For backwards compatibility, this PR keeps the `kubelet_cache_pods_duration` config option for users who still want to set it to reduce pulls from the kubelet.

### Describe how you validated your changes

CI


Co-authored-by: david.ortiz <david.ortiz@datadoghq.com>