[EBPF-910] gpu: exclude datadog containers from GPU device mapping (#43225)
### What does this PR do?

This PR ensures that Datadog containers are excluded from the GPU mapping assignments.

### Motivation

Agent containers can have access to the GPU, and in some cases such as Docker, they can use it in the same way as regular workloads. This would make the agent show up as a consumer of GPUs in the UI, which could cause confusion.

### Describe how you validated your changes

Unit tests included.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>