Allow running pulumi up in test/e2e-framework and fix import check (#43414)
### What does this PR do?

Change back the root package of test/e2e-framework to a `main` package. This is needed because we need that package to be an executable to have the `dda inv create-*` tasks working, as well as calling pulumi up directly.

I updated the check-mod-tidy task because the task checks that there is no issue by importing all the modules and running go mod tidy. It was not working with the import of root package of `test/e2e-framework` because that was an executable. To workaround that issue I added a directive to modules.yaml that allows to specify a package that is importable in the module, when the root is not. It should make the test behaves correctly as we are mostly interested in making sure there are no incompatibilities in modules go.mod, so importing any package of the module should be fine.

### Motivation

Fix the local infra creation tasks.

### Describe how you validated your changes

Ran the task locally to make sure it works, as well as the `go_mod_tidy_check` job.

### Additional Notes


Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>