[EBPF-876] gpu: fix e2e tests (#42588)
### What does this PR do?

This PR fixes GPU e2e tests which were previously broken. The issue was that the container image we use for the workload changed and was no longer defining `NVIDIA_VISIBLE_DEVICES`, which means that it no longer could access GPU devices. This issue was hard to detect because the error message shown only mentioned a mismatch between CUDA and driver versions, nothing about the devices themselves not being accessible.

There was also a test failure introduced in a previous PR (https://github.com/DataDog/datadog-agent/pull/42524), where the format of workloadmeta GPUs changed (the source is now `nvml` instead of  `runtime`) and that wasn't taken into account in the tests.

Additionally, this PR includes some fixes to make debugging of the e2e tests easier:

- A new `CheckWorkloadError` method is developed so that the tests can check that the workload container did not fail.
- The provisioner validates that the CUDA version installed is the one expected.
- The provisioner validates that the driver is installed and running.

### Motivation

Fix disabled tests and re-enable them.

### Describe how you validated your changes

Tested in a local run, retried the job in CI multiple times to ensure it works.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>