[ABLD-302] Start switching pkg_install usage to know the full path of where to install things.
- Restore the ability to have an addition prefix on so_symlink.
- Change --dest_dir={install_dir}/embedded to --dest_dir={install_dir}
- Add prefix='embeded' to thigs we push

As we start installing more, we will have some pieces that go to embedded, and others that do not.
It will be easier to understand if all the install instances use the single destdir.

But also, we would like to be able to get dest dir as a build flag
https://datadoghq.atlassian.net/browse/ABLD-302

The need for that is to automate build time genrules where we must brand the dest_dir into a binary.
At that point we can also use the command line dest_dir in the pkg_install command,
eliminating the need to use --dest_dir on the command line.