[Serverless-Init] Cloud Run Jobs Inferred Span (#43371)
### What does this PR do?

Create an inferred span for the duration of the task execution.
<img width="600" height="744" alt="Screenshot 2025-11-21 at 4 34 44 PM" src="https://github.com/user-attachments/assets/2ff7a596-0233-4d71-90d1-36215e42ec8e" />
[View span](https://ddserverless.datadoghq.com/apm/trace/1313316933801504975?spanID=890698588823561473&timeHint=1763759967411.9866&trace=1313316933801504975890698588823561473)

And it correctly reparents spans received from the tracer (custom, autoinstrumentation, etc.)
<img width="600" height="172" alt="Screenshot 2025-11-21 at 6 45 31 PM" src="https://github.com/user-attachments/assets/ec54d798-7e35-4e9d-b14f-61075997bdd9" />

This can be disabled with the env var `DD_APM_ENABLED=false`

### Motivation

So traces work OOTB, no custom span creation is required from the user. This also follows the convention set by Lambda.

As an added benefit, we can get rid of the high cardinality tags on the existing enhanced metrics. Instead, we can just rely on this trace existing to display the duration of each task for a specific execution ID on the frontend. (This will be done in a future PR, and it requires frontend changes)

### Describe how you validated your changes

Testing manually with a real Cloud Run Job

### Additional Notes

While it's not ideal to have APM logic in the agent, there is no other way to do this. In Lambda, the tracers can instrument the Lambda handler function. But in Cloud Run Jobs, there is no such entry point to instrument. The tracer only knows it's in a container, but not that it's in a Cloud Run Job.

We have a precedent for creating spans in the agent (in the [Lambda Extension](https://github.com/DataDog/datadog-lambda-extension/)), and there were no issues there.


Co-authored-by: nicholas.hulston <nicholas.hulston@datadoghq.com>