usm: native-tls: Fix a leak while terminating the context
We used tup_from_ssl_ctx method during the termination and cleanup of the context, the method can trigger the
mechanism to map again ctx to the socket during the next tcp_sendmsg which can occur while sending fin/ack.

Furthermore, we pulled an address from a map, then deleted the entry from the map, and then used the address
as a valid address. It creates an ambigious behavior, as the content pointed by the address can be overriden
by another call in a different cpu, leading to a malformed state or using incorrect tuple.

In the fix, we ensure we are in a read only mode (directly pulling from the map), we copy locally the conn_tuple
and then deleting the entry from the map