[CWS/CSPM] simplify log context creation (#43491)
### What does this PR do?

When creating the logs context (the endpoint structure containing the intake url, the api key, some endpoint configuration etc), we need to make sure we are actually using the correct endpoint (i.e. pointing to the runtime security or compliance intake, with the correct path and the correct api key). Falling back on the logs endpoints (i.e. either configured through the logs config, or even sending the data to the logs intake) doesn't achieve this.

This PR removes the fallback mechanism where we would fall back on the logs config/intake if we were failing to build the expected endpoint configuration.

This also means that we can now correctly report endpoint configuration errors to the users (as they would most likely expect).

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: paul.cacheux <paul.cacheux@datadoghq.com>