[EBPF] gpu: use workloads IDs to generate tags (#43336)
### What does this PR do?

This PR refactors the GPU check tagging code to use workload IDs to generate tags. The collectors that emit metrics associated to a specific workload (e.g., per-process metrics) will now add the `EntityID` of that workload instead of creating the tags themselves. If a metric has no workloads associated, the code associates it with all the containers assigned to the corresponding GPU.

To do this, we modify the `containerTagCache` to be a generic worload tag cache, which returns tags for a given workloadmeta entity.

- Retrieving the tags for each entity. This is done by calling the tagger with the proper arguments (e.g., different cardinality depending on the container runtime), by adding non-standard tags (e.g., processes do not have the `pid` or `nspid` tag in the tagger, and we want to keep it that way) and by relying on fallback mechanisms if workloadmeta is not available.
- Caching tags for entities, so that we are not invoking the tagger every single time.
- Caching stale data for entities, so that we can emit tags even if the entity has disappeared already (e.g., we might still want to add tags for processes that are no longer running when we collect the data).

The underlying structure is a LRU cache, to avoid growing indefinitely and to simplify the eviction logic. Cache size is configurable, defaulting to 1024 entries.

Part of the code for the cache has been taken from the `nsPidCache` implemented previously, which accomplished a similar thing except it was only scoped for namespace PIDs.

### Motivation

This is necessary as part of https://datadoghq.atlassian.net/browse/EBPF-911. Right now, the collectors tag the metrics with any process-related tags (`pid`, `nspid`) while the check itself adds the tags from the container assigned to the GPU to every metric for that GPU. This the correct behaviour as long as only one container could be assigned to one GPU. However, with multiple containers being assigned to a GPU, a process-level metric (such as `gpu.process.memory.usage`) only should be tagged with the process and container tags relevant to that process, not with every possible container that can use that GPU.

This refactor generalizes that logic, and separates the responsibilities for adding tags from the workload assignment.

### Describe how you validated your changes

Unit tests written for the tag cache. Also manually ran the agent to check that workloads are getting the expected tags.

This PR will also require validation during the RCs.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>