Add helpers to opentelemetry-mapping-go (#43196)
### What does this PR do?
 
Add helpers to mapping-go to compute fields from otel attributes. These will be used for trace metrics computations in stats backend (followup to https://github.com/DataDog/dd-go/pull/198508). Logic will be different than the methods in that PR: for this version, we will no longer use the `datadog.*` namespace, and also will only be checking for attributes in the location that aligns with the OTel spec.



Co-authored-by: ibraheem.aboulnaga <ibraheem.aboulnaga@datadoghq.com>