Remove UnmarshalKey from config API, call mapstructure.Decode (#43417)
### What does this PR do?

UnmarshalKey is deprecated on the config interface. Instead, use our own version in pkg/config/structure. Call mapstructure.Decode directly, copying using GetViperCombine for the viper version.

### Motivation

After performance testing, we found that calling mapstructure.Decode directly seems to run faster. This used to not work in Viper, because it didn't merge layers of the config, but now that we have GetViperCombine, that isn't the case any more. In theory, it should be possible to make the reflection based UnmarshalKey run faster for nodetreemodel, since it doesn't need to copy as much, but doing it this way is fine for now.

### Describe how you validated your changes

Unit tests.

### Additional Notes


Co-authored-by: dustin.long <dustin.long@datadoghq.com>