[AGENTRUN-548] Add slog based implementation replacing seelog (#43297)
### What does this PR do?
Add an slog based implementation of `LoggerInterface` replacing seelog, controlled by the `log_use_slog` configuration (disabled by default).

### Motivation
Eventually replace seelog with slog-based logger entirely.

Seelog is unmaintained for years, and doesn't provide many features that we need, forcing us to add layers and layers on top of it, making the code very complex.
This migration will hopefully allow us to significantly simplify the code later, and make it easier to add features.

### Describe how you validated your changes
- Added many unit tests for the new code.
- Some existing e2e tests cover logging (eg. flares check that they contain the log file)
- Some manual high level testing (running with common/json format and checking that the logs are the same, checking the logs are written to file, etc)
- TODO: manual testing for the more complex features (syslog)

Once merged we will test the new implementation internally.

### Additional Notes
Replaces https://github.com/DataDog/datadog-agent/pull/39139 (and then https://github.com/DataDog/datadog-agent/pull/42846 as well), reusing a lot of the new code.
Most of the new code was introduced in previous small PRs.

Some logic is somewhat duplicated (eg. log formats) between the old seelog code and the new slog code, but both should live close-by (eg. old and new formatters are in the same file).

We could use libraries for some of the logic, but it would be difficult to be sure we're 100% equivalent to the existing implementation, which is a requirement (in particular for the testing phase, as it makes the transition easier)
I do want to take a look at some of those later though.
- https://github.com/samber/slog-multi
- https://github.com/samber/slog-syslog
- https://github.com/samber/slog-formatter
- https://github.com/samber/slog-sampling
- https://github.com/samber/slog-zap

Co-authored-by: pierre.gimalac <pierre.gimalac@datadoghq.com>