[EBPF] gpu: cache tags in case of error (#43578)
### What does this PR do?

Changes the workload tag cache so that any tags built are stored in the cache for subsequent runs. The `TestGetWorkloadTags` has also been fixed to account for this case.

The e2e test has also been fixed to ensure that the `pid:` tag is always present on core usage metrics.

### Motivation

Without this fix, there would be an inconsistency where in the first call to get tags for a missing process we would get some default tags (pid/nspid) but then in subsequent ones we wouldn't get any. It also caused #incident-46337: when `gpu.process.core.usage` was the metric that got the `pid` tags, then `gpu.core.limit` wouldn't get them and therefore we did not find it and failed the test.

### Describe how you validated your changes

Unit tests added to ensure this behavior is covered and working.

### Additional Notes


Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>