[ABLD-184] Copy licenses for Bazel built deps into product build tree. (#43507)
Add capability to export the license texts to /opt/datadog-agent/LICENSES.

- Finds the used licenses of a top level target. Until the time we can build the agent binaries with Bazel, //deps:all_deps is a proxy for that.
- Call from agent-finalize.

This brings us to parity with what the agent artifacts include now. That meets the immediate goal of being able to delete omnibus steps. I'm working with the open source team on improvements to the overall strategy, but that is not a blocker to a bug-for-bug replacement today.

The build file is ugly.  It could be better if we improved rules_pkg, but that is for later.

- Grab the build .deb file from CI
- Crack it open and look.


Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>