[EBPF] gpu: use common process monitor in event consumer (#43551)
### What does this PR do?

Replaces the use of the deprecated process monitor from `pkg/process/monitor` to watch for process exits, uses instead the one that's already passed to the uprobe attacher.

### Motivation

Uniform usage of process monitors. This will also solve https://datadoghq.atlassian.net/browse/EBPF-850, as that's due to the design of the process monitor in `pkg/process/monitor` where callbacks can still be executed after the unsubscribe function is called. In this case, we would call the unsubscribe function and then close the channel, and the callback would still execute in some rare cases.

### Describe how you validated your changes

Unit tests passing.

### Additional Notes

Co-authored-by: guillermo.julian <guillermo.julian@datadoghq.com>