[ABLD-239] Pull in gstatus as a bazel reference. (#43497)
There is a slight departure from the way we have built most of the C, linked in dependencies
 
- Install the license file directly with the rest of the install. 
- lift the pkg_install up to datadog-agent-integrations-py3-dependencies.rb
- delete gstatus.rb

Installing the license whenever we install the binary is the safest choice. Nothing depends on gstatus as a direct dependency, so we would never catch the license until the time when we build the entire distribution image with bazel.  Doing it this way now prevents losing the license.

Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>