Add missing dependencies when updating the release.json file (#43625)
### What does this PR do?

Add missing dependencies when updating the release.json file

It also sorts the key alphabetically (other bots are doing the same) so we stay consistent

### Motivation

When new keys are added to the dependencies we need to explicitly set them in the function. If we forget, we loose them. This PR fixes that

### Describe how you validated your changes

### Additional Notes


Co-authored-by: florent.clarret <florent.clarret@datadoghq.com>