[NDMII-3715] Fix profile not refreshed when the profile is an inline profile (#43572)
### What does this PR do?

https://datadoghq.atlassian.net/browse/AGENT-15023

The support ticket above has a customer that needs to restart their Agent to pick-up a metric.

This is because [this PR](https://github.com/DataDog/datadog-agent/pull/39225) introduced the logic to delete the OIDs to query that the device do not report, and the OIDs to query were refreshed every 10min again to pick-up new OIDs. However the logic to refresh every 10min only applied for OIDs defined in profiles and not the inline OIDs defined in the instance config (which is the case in the support ticket). This PR aims to fix this by refreshing the profile for whatever way the user defined their OIDs.

### Motivation

### Describe how you validated your changes

### Additional Notes


Co-authored-by: pierre.lin <pierre.lin@datadoghq.com>