[ABLD-46] Create individual files for ship source offers.
This change extends licenses_csv to be able to create indivudual offer files for each package. This is in addition to the previous capability to emit all the offers into a single file. Single files are the way we have previously done this, so the current default is that legacy technique. As a low priority effort, I'll be advocating to condense them to a single file, so both output styles co-exist.