[ABLD-46] Create individual files for ship source offers. (#43689)
This change extends licenses_csv to be able to create individual offer files for each package. This is in addition to the previous capability to emit all the offers into a single file. Single files are the way we have previously done this, so the current default is that legacy technique. As a low priority effort, I'll be advocating to condense them to a single file, so both output styles co-exist.

### Sample
```
$ bazel run -- //compliance:install_licenses --destdir=/tmp/xi
$ find /tmp/xi
/tmp/xi
/tmp/xi/LICENSES
/tmp/xi/LICENSES/xz-COPYING.0BSD
/tmp/xi/LICENSES/-COPYING.LGPL
/tmp/xi/LICENSES/zlib-LICENSE
/tmp/xi/LICENSES/libffi-LICENSE
/tmp/xi/LICENSES/bzip2-LICENSE
/tmp/xi/LICENSES/nghttp2-COPYING
/tmp/xi/sources
/tmp/xi/sources/acl
/tmp/xi/sources/acl/offer.txt
/tmp/xi/sources/attr
/tmp/xi/sources/attr/offer.txt
$ cat /tmp/xi/sources/attr/offer.txt
This package ships attr@2.5.1
Contact Datadog Support (http://docs.datadoghq.com/help/) to request the sources of this software.
```

Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>