fix bug in calculating name for copies of license files (#43699)
Fix bug in calculating name for copies of license files when the pacage is defined with a PURL rather than a bazel repository.

The license may be associated with old style license rules, giving us something like `@libz//:LICENSE` or with a PURL, giving us something like `pkg:/some/stuff/libz@1.2.3?url=more-stuff`.  The PURL should be prefered, but the extraction for the name was not right, so we were getting an empty string.

Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>