privileged logs: Fix handling of unsearchable directories
In any directory in the path of the log file does not have executable
permissions, it does not get tailed since a stat(2) call from the file
launcher fails and the logs agent never gets to the point where it calls
into system-probe.  Fix this by handling that stat(2) call too in
system-probe if privileged logs is enabled.