fix(ssi): resolve bug with disabled label in enabled namespace (#43543)
### What does this PR do?
This commit resolves a bug where a pod with a disabled label in an
enabled namespace would still receive injection.

### Motivation
When working through a support case, I noticed that we did not respect
the disabled label in an enabled namespace.

### Describe how you validated your changes
The additional unit tests.

### Additional Notes
This entire webhook is scheduled to be refactored. When doing so, we
should consider a more sophisticated version of the `annotationResult`:
<img width="1062" height="481" alt="Screenshot 2025-11-26 at 11 56
36 AM"
src="https://github.com/user-attachments/assets/94df3a41-0bf6-481b-9096-4ef1c2e5ee87"
/>

---------

Co-authored-by: domalessi <111786334+domalessi@users.noreply.github.com>