add peds-based docker proxy filter for direct send (#43438)
### What does this PR do?

Adds a "process event data stream" (PEDS) based docker proxy filter. It is used to filter out connections that related to `docker-proxy` processes that are not useful for customers to see.

### Motivation

This will be used in CNM when sending data directly to the backend. 

### Describe how you validated your changes

Added tests.

### Additional Notes

This is a "port" of existing behavior in `pkg/process/metadata/parser/dockerproxy.go` used when CNM data is flowing through `process-agent`. It had to be adapted to a different data source for processes AND different data types for the network connections data.


Co-authored-by: bryce.kahle <bryce.kahle@datadoghq.com>