Start switching pkg_install usage to know the full path of where to install things. (#43487)
[ABLD-302] Start switching pkg_install usage to know the full path of where to install things.

- Restore the ability to have an addition prefix on so_symlink.
- Change --dest_dir={install_dir}/embedded to --dest_dir={install_dir}
- Add prefix='embedded' to things we push

This did some of the libraries. I can do the rest in a followup or in this PR.

# Rationale

As we start installing more, we will have some pieces that go to embedded, and others that do not. It will be easier to understand if all the install instances use the single destdir. Eventually, we can create targets like //pkg/install_dir:install, which is a big pkg_install of all the dependencies and config scripts. Then the whole thing gets built  with a single `bazel run -- //opt/datadog/lib:install --destdir='#{install_dir}'.   When we get multiple remote executors, it can all dispatch in parallel.

The next release of rules_pkg will allow destdir to come from a build flag. The need for that is to automate build time genrules where we must brand the dest_dir into a binary. At that point we can also use the command line dest_dir in the pkg_install command, eliminating the need to use --dest_dir on the command lines in .rb files.



[ABLD-302]: https://datadoghq.atlassian.net/browse/ABLD-302?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

Co-authored-by: tony.aiuto <tony.aiuto@datadoghq.com>