Move code from datadog-agent to e2e-framework, reducing duplicate (#43453)
### What does this PR do?

Move a lot of code that was leaving in datadog-agent inside test/e2e-framework to make it a self-sufficient module that define environments that can be used both for e2e tests and for local qa

That PR moves the `new-e2e/pkg` folder inside the e2e-test framework. That way other repo that need to use the tooling to execute test just need to import `test/e2e-framework` and not two modules `test-infra-definitions` + `test/new-e2e` as of today.
I moved all the AWS provisioners run function inside scenarios, that can be reused either with the provisioner in the e2e tests or through pulumi up / `create-*` tasks.
There some code about Azure and GCP provisioner that still need to be moved, will be done in a separate PR to makes it easier to review

### Motivation

Remove duplicate to ensure local environment are as well maintained as e2e tests environments

### Describe how you validated your changes

All tests being green in the CI


### Additional Notes


NOTES FOR REVIEWER:

The PR should update import and wrap most of the provisioner options inside `WithRunOptions`, if you observe anything else on your test please comment it, that's not expected.

Co-authored-by: kevin.fairise <kevin.fairise@datadoghq.com>