[APMLP-730] Change the typing of `admission_controller.auto_instrumentation.default_dd_registries` config from map to list (#43691)
### What does this PR do?
As the title says, this PR changes the typing of the `admission_controller.auto_instrumentation.default_dd_registries` config from a `map[string]any` to `[]string`.

### Motivation
This change aims to make the configuration more intuitive for users (not customers, since this is an internal-only config). The original implementation was optimizing for the ImageResolver, so that it could directly use the hash map/set from the config to determine what valid Datadog registries were. However, this was a suboptimal/unintuitive experience for users having to set a map for the config value, when what really mattered were the string key values. 

### Describe how you validated your changes
Validated existing tests passed when the typing changed.

### Additional Notes


Co-authored-by: erika.yasuda <erika.yasuda@datadoghq.com>