[CWS] Use modulo for flags approvers (#43825)
### What does this PR do?

This PR computes the modulo of the uint64 value used for filtering with approvers, ensuring that the value remains non-zero after the bit-shift operation.

### Motivation

When using flag approvers, we perform a shift operation (1 << x). If the value we want to filter on is greater than or equal to 64, the result becomes 0 due to the overflow of the 64-bit shift.
By applying a modulo operation before shifting, we ensure that the computed filter is always valid and non-zero.

Upsides:
Some values were not added to the approvers, which resulted in a significant increase in the number of events sent to user space. This fix ensures that these values are now correctly handled, reducing unnecessary event volume.

Downside:
This change may allow some events with a value ≥ 64 to be sent to user space even if they are not strictly needed. However, this downside should be negligible compared to the benefits provided by the fix.

Co-authored-by: theo.putegnat <theo.putegnat@datadoghq.com>