dyninst/procsubscribe: expose debug info for tracked processes
This patch adds a debug endpoint /dynamic_instrumentation/debug/goprocs
(to be accessed through `system-probe debug dynamic_instrumentation
goprocs) that lists information about Go processes that are "tracked"
(i.e. the remote config module in the core agent was told to route
updates for the respective runtime ID to the system-probe), as well as
other processes that have been detected on the box. The two sets should
be the same, but we appear to have a bug that I'm trying to track down.